Updated TryParseData function
Down the road of my tests I've noticed that the problems of converting data on some formats was due to the usage of a '.' while C# expected ','

This behaviour might change from culture to culture.

Worksheet and the other vars were then moved outside TryParseDate as I noticed that if there is no DataType then we must guide ourselves by formatIds value.