Updated GetCellValue
Updated this function to be able to return an object so that less casts and data type convertions are needed.

Now I've added a switch to control data parses based on formatId if the cell doesn't contain DataType.

For more info on how to parse data from these formatIds I've looked into OpenXML docs but being too extensive I couldn't find much, however Microsoft does provide documentation on OpenXML and it can be found at: https://docs.microsoft.com/en-us/dotnet/api/documentformat.openxml.spreadsheet.numberingformat?view=openxml-2.8.1

Finally I've updated the return value of the Boolean case to return a true of false and thus avoiding parsing to string and viceversa to read and save boolean values.

Depending on wether cell.CellFormula is null (meaning it has the name implies that a formula exists) or not we send either cell.CellValue.Text or cell.InnerText

The 1st contains the result of the formula execution and the former contains the value we want if the cell value is static